requirements.txt
# SpotBot dependencies
 
# Discord bot framework
discord.py>=2.3.0
 
# Spotify API wrapper
spotipy>=2.23.0
 
# Environment variable loading
python-dotenv>=1.0.0

# pyest
pytest>=8.0.0
pytest-asyncio>=0.23.0